M6 Aufgabe 1c-6 & 2 3)-4)

+++ emensa/models/Meal.php
@@ -18,7 +18,7 @@ class Meal
     public float $priceIntern;
     public float $priceExtern;
     public array $allergens;
-    public int|null $rating;
+    public float|null $rating;
 
     public function __construct(array $mealData)
     {
